Reimplement ExifTagIterator
Compiles, which is nice, and uses an iterator for the section offsets
which is modified while the tags if IFD0 are iterated to find the offsets
off the other sections.

Also, reimplements flat_map because it has to deal with errors and
the borrow to mutate the SectionOffsetIterator from the flat_map closure
was too complicated.