reduce memory footprint of Cursor
instead of keeping an offset, advance the slice as the cursor moves.
the only disadvantage of this is that errors can indicate the location
of the cursor anymore. Those errors are not recoverable anyway, but may be
useful for diagnosing bugs... need to think if it's worthwhile, but think it is.