Fix calculation of IFD1 offset location
The IFD1 offset comes right after the IFD0 section, but
the we weren't taking into account that IFD0 also has an offset, and that IFD0 is bigger than just the size of its tags (it has 2 bytes for the length in front as well) when jumping to that location, so we ended up with a wrong IFD1 offset that tried to jump to a random offset and failed in most cases.

This brings back the byte_size method to the section iterator because a section has the knowledge of what parts are in a section. Also it stores the offset of the current section while iterating it, so we can use the IFD0 offset to calculate the location of the IFD1 offset after finishing with IFD0.